util 함수 코드 위치 이동
